+++ src/shared/src/types/decimal.rs
@@ -6,20 +6,20 @@ use serde::Serialize;
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 pub struct DecimalValue {
     /// The numeric value (mantissa).
-    pub value: u64,
+    pub value: u128,
     /// The number of decimal places (exponent).
     pub decimals: u8,
 }
 
 impl DecimalValue {
     /// Creates a new DecimalValue.
-    pub fn new(value: u64, decimals: u8) -> Self {
+    pub fn new(value: u128, decimals: u8) -> Self {
         Self { value, decimals }
     }
 }
 
-impl From<(u64, u8)> for DecimalValue {
-    fn from((value, decimals): (u64, u8)) -> Self {
+impl From<(u128, u8)> for DecimalValue {
+    fn from((value, decimals): (u128, u8)) -> Self {
         Self::new(value, decimals)
     }
 }
